updated version, and requirements. Added delete bedset to cli
