+++ requirements/requirements-all.txt
@@ -4,8 +4,8 @@ peppy>=0.40.1
 yacman>=0.8.4
 requests>=2.28.2
 piper>=v0.14.0
-# bbconf>=0.6.0
-bbconf @ git+https://github.com/databio/bbconf.git@thinking#egg=bbconf
+bbconf>=0.5.0
+# bbconf @ git+https://github.com/databio/bbconf.git@dev#egg=bbconf
 refgenconf>=0.12.2
 pandas>=1.5.3
 ubiquerg>=0.6.2
